added remove duplicated minuses
